Don't use a full completion snippet when completing a method with existing args or block
Previously when completing a method name that already had an argument list or a block attached, our snippet would do double duty and re-add a whole new set of them.

With this change we first check that the send site we are completing doesn't already have those specified and in that case simply focus on completing the name.